refactor: move file manager module to separate directory
